python_scripts/requirements.txt
#Deep Learning Framework
torch>=2.0.0            #PyTorch: base para treinar/redes neurais
torchvision>=0.15.0     #Pacotes extras para visão computacional (datasets, transforms, modelos)

#Modelos pré-treinados / arquiteturas
timm>=0.9.2             #Coleção de modelos de visão (ViT, EfficientNet, etc.)
transformers>=4.33.0    #Hugging Face Transformers (ConvNeXt, ViT, etc.)
pretrainedmodels @ git+https://github.com/Cadene/pretrained-models.pytorch.git
                        #Pacote antigo com ResNet, Inception, etc.
                        #Mantido via Git, instalar por último pois pode demorar

#Segmentação / Remoção de fundo
rembg>=2.0.50           #Biblioteca para remover fundo de imagens (usa deep learning)
onnxruntime>=1.16.0     #Backend de execução usado pelo rembg (para rodar modelos ONNX)

#Utilitários gerais
pillow>=10.0.0          #Pillow (PIL): abrir, manipular e salvar imagens
numpy>=1.24.0           #Numpy: operações com arrays e cálculos
opencv-python>=4.8.0    #OpenCV: processamento de imagens tradicional (segmentação, filtros)
pandas>=2.0.0           #Pandas: manipulação de tabelas/resultados
albumentations>=1.3.0   #Data augmentation avançado para imagens
protobuf>=4.25.0        #Compatibilidade ONNX / Transformers
requests>=2.31.0        #Requisições HTTP (Hugging Face, downloads de modelos)
tqdm                    #Barra de progresso para loops e treinamentos
scipy                   #Operações matemáticas avançadas
natsort                 #Ordenação “natural” de strings, útil para nomes de arquivos

#Machine Learning clássico
scikit-learn>=1.3.0     #Classificadores (SVM, RandomForest), métricas, validação